chore: deprecate server.ts
